fix: resolve pre-push gate findings (probing budget, dead seniorityRank, DTO fixtures)
