+++ web/src/__tests__/pages/ProviderDetailPage.test.tsx
@@ -38,6 +38,8 @@ function makeProvider(name: string): ProviderWithName {
           supports_vision: false,
           supports_reasoning: false,
           max_output_tokens: null,
+          parameter_count: null,
+          cost_tier: null,
           family: null,
           generation: null,
           release_date: null,
