+++ web/src/__tests__/pages/setup/AgentsStep.test.tsx
@@ -42,6 +42,8 @@ function model(
       supports_vision: false,
       supports_reasoning: false,
       max_output_tokens: null,
+      parameter_count: null,
+      cost_tier: null,
       family: null,
       generation: null,
       release_date: null,
