+++ web/src/__tests__/utils/providers.test.ts
@@ -168,6 +168,8 @@ describe("sortProviders", () => {
             supports_vision: false,
             supports_reasoning: false,
             max_output_tokens: null,
+            parameter_count: null,
+            cost_tier: null,
             family: null,
             generation: null,
             release_date: null,
@@ -193,6 +195,8 @@ describe("sortProviders", () => {
             supports_vision: false,
             supports_reasoning: false,
             max_output_tokens: null,
+            parameter_count: null,
+            cost_tier: null,
             family: null,
             generation: null,
             release_date: null,
@@ -213,6 +217,8 @@ describe("sortProviders", () => {
             supports_vision: false,
             supports_reasoning: false,
             max_output_tokens: null,
+            parameter_count: null,
+            cost_tier: null,
             family: null,
             generation: null,
             release_date: null,
