+++ web/src/__tests__/utils/setup-validation.test.ts
@@ -56,6 +56,8 @@ const makeModel = (
     supports_vision: false,
     supports_reasoning: false,
     max_output_tokens: null,
+    parameter_count: null,
+    cost_tier: null,
     family: null,
     generation: null,
     release_date: null,
