+++ web/src/pages/providers/ProviderCard.stories.tsx
@@ -23,6 +23,8 @@ const baseProvider: ProviderWithName = {
         supports_vision: false,
         supports_reasoning: false,
         max_output_tokens: null,
+        parameter_count: null,
+        cost_tier: null,
         family: null,
         generation: null,
         release_date: null,
@@ -43,6 +45,8 @@ const baseProvider: ProviderWithName = {
         supports_vision: false,
         supports_reasoning: false,
         max_output_tokens: null,
+        parameter_count: null,
+        cost_tier: null,
         family: null,
         generation: null,
         release_date: null,
